requirements.txt
meraki
azure-mgmt-network
azure-identity